skills/raster-to-vector/scripts/pyproject.toml
[project]
name = "r2v"
version = "0.1.0"
requires-python = ">=3.12,<3.13"
dependencies = [
  "numpy>=1.26",
  "opencv-python>=4.13.0.92",
  "paddlepaddle>=3.0",
  "paddleocr>=3.5,<4",
  "pillow>=12.2.0",
  "vtracer>=0.6.15",
  "mlx-vlm>=0.4.4; sys_platform == 'darwin' and platform_machine == 'arm64'",
]

[tool.uv]
package = false
